Change default behaviour for wrong option values
Before, if FindBounce was given wrong option value,
a message was issued and default value for option was used.
Now a message is still issued, but the function returns $Failed.
This behaviour avoids unpleasant surprises if a user uses Quiet
on whole function or supresses messages in some other way.
Tests for this behaviour are added as well.