Add switch to check the results only at the end of the iterations
This make more stable the package. This switch is True only in the
last iteration of the path. If V1+T1 is smaller than zero it should
return $Failed, but if there are more iteration of the path to be done,
it waits and make this check at the end, since the results could be
improved after each iteration, as it actually happens.